// interfaceApi/models/InterfaceInformation.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type InterfaceInformation = {
  createTime?: string;
  id?: number;
  introduce?: string;
  isDelete?: number;
  maven?: string;
  method?: string;
  name?: string;
  paramName?: string;
  requestHead?: string;
  responseHead?: string;
  sdkURL?: string;
  status?: number;
  updateTime?: string;
  url?: string;
  userId?: number;
};