// interfaceApi/models/InterfaceInformationAddRequest.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type InterfaceInformationAddRequest = {
  createTime?: string;
  id?: number;
  isDelete?: number;
  maven?: string;
  method?: string;
  name?: string;
  requestHead?: string;
  responseHead?: string;
  status?: number;
  updateTime?: string;
  url?: string;
  userId?: number;
};