// interfaceApi/models/InterfaceInformationUpdateRequest.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type InterfaceInformationUpdateRequest = {
  id?: number;
  method?: string;
  name?: string;
  requestHead?: string;
  responseHead?: string;
  status?: number;
  url?: string;
  userId?: number;
};